+++ CHANGES.rst
@@ -6,7 +6,7 @@ Changelog of lizard-fewsjdbc
 1.31.4 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Values can be localized too when using the api.
 
 
 1.31.3 (2012-10-05)